+++ src/snowflake/cli/api/project/schemas/native_app/package.py
@@ -59,5 +59,5 @@ class PackageV11(Package):
     # Templated defaults only supported in v1.1+
     name: Optional[str] = IdentifierField(
         title="Name of the application package created when you run the snow app run command",
-        default="<% fn.id_concat(ctx.native_app.name, '_pkg_', fn.clean_id(fn.get_username())) %>",
+        default="<% fn.id_concat(ctx.native_app.name, '_pkg_', fn.clean_id(fn.get_username('unknown_user'))) %>",
     )
