+++ tests/api/utils/test_definition_rendering.py
@@ -835,303 +835,3 @@ def test_defaults_native_app_pkg_name(
     env = project_context.get("ctx", {}).get("env", {})
     assert env.get("app_reference") == expected_app_name
     assert env.get("pkg_reference") == expected_pkg_name
-
-
-def test_template_unknown_function():
-    definition = {
-        "definition_version": "1.1",
-        "env": {
-            "value": "<% fn.unknown_func('hello') %>",
-        },
-    }
-
-    with pytest.raises(InvalidTemplate) as err:
-        render_definition_template(definition, {})
-
-    assert "Could not find template variable fn.unknown_func" in err.value.message
-
-
-@pytest.mark.parametrize(
-    "input_list, expected_output",
-    [
-        # test concatenate a constant with a variable -> quoted
-        (["'first_'", "ctx.definition_version"], '"first_1.1"'),
-        # test concatenate valid unquoted values  -> non-quoted
-        (["'first_'", "'second'"], "first_second"),
-        # test concatenate unquoted ids with unsafe chars -> quoted
-        (["'first.'", "'second'"], '"first.second"'),
-        # all safe chars, one with quoted id -> quoted
-        (["'first_'", "'second_'", "'\"third\"'"], '"first_second_third"'),
-        # one word, unsafe chars -> quoted
-        (["'first.'"], '"first."'),
-        # one word, safe chars -> non-quoted
-        (["'first'"], "first"),
-    ],
-)
-def test_id_concat_with_valid_values(input_list, expected_output):
-    input_list_str = ", ".join(input_list)
-    definition = {
-        "definition_version": "1.1",
-        "env": {
-            "value": f"<% fn.id_concat({input_list_str}) %>",
-        },
-    }
-
-    result = render_definition_template(definition, {}).project_context
-    env = result.get("ctx", {}).get("env", {})
-    assert env.get("value") == expected_output
-
-
-def test_id_concat_with_no_args():
-    definition = {
-        "definition_version": "1.1",
-        "env": {
-            "value": f"<% fn.id_concat() %>",
-        },
-    }
-
-    with pytest.raises(InvalidTemplate) as err:
-        render_definition_template(definition, {})
-
-    assert "id_concat requires at least 1 argument(s)" in err.value.message
-
-
-def test_id_concat_with_non_string_arg():
-    definition = {
-        "definition_version": "1.1",
-        "env": {
-            "value": "<% fn.id_concat(123) %>",
-        },
-    }
-
-    with pytest.raises(InvalidTemplate) as err:
-        render_definition_template(definition, {})
-
-    assert "id_concat only accepts String values" in err.value.message
-
-
-@pytest.mark.parametrize(
-    "input_val, expected_output",
-    [
-        # unquoted safe -> unchanged
-        ("first", "first"),
-        # unquoted unsafe -> unchanged
-        ("first.second", "first.second"),
-        # looks like quoted but invalid -> unchanged
-        ('"first"second"', '"first"second"'),
-        # valid quoted -> unquoted
-        ('"first""second"', 'first"second'),
-    ],
-)
-def test_id_to_str_valid_values(input_val, expected_output):
-    definition = {
-        "definition_version": "1.1",
-        "env": {
-            "input_value": input_val,
-            "output_value": "<% fn.id_to_str(ctx.env.input_value) %>",
-        },
-    }
-
-    result = render_definition_template(definition, {}).project_context
-    env = result.get("ctx", {}).get("env", {})
-    assert env.get("output_value") == expected_output
-
-
-def test_id_to_str_with_no_args():
-    definition = {
-        "definition_version": "1.1",
-        "env": {
-            "value": "<% fn.id_to_str() %>",
-        },
-    }
-
-    with pytest.raises(InvalidTemplate) as err:
-        render_definition_template(definition, {})
-
-    assert "id_to_str requires at least 1 argument(s)" in err.value.message
-
-
-def test_id_to_str_with_two_args():
-    definition = {
-        "definition_version": "1.1",
-        "env": {
-            "value": "<% fn.id_to_str('a', 'b') %>",
-        },
-    }
-
-    with pytest.raises(InvalidTemplate) as err:
-        render_definition_template(definition, {})
-
-    assert "id_to_str supports at most 1 argument(s)" in err.value.message
-
-
-def test_id_to_str_with_non_string_arg():
-    definition = {
-        "definition_version": "1.1",
-        "env": {
-            "value": "<% fn.id_to_str(123) %>",
-        },
-    }
-
-    with pytest.raises(InvalidTemplate) as err:
-        render_definition_template(definition, {})
-
-    assert "id_to_str only accepts String values" in err.value.message
-
-
-@pytest.mark.parametrize(
-    "input_val, expected_output",
-    [
-        # unquoted safe -> unchanged
-        ("first", "first"),
-        # unquoted unsafe -> quoted
-        ("first.second", '"first.second"'),
-        # looks like quoted but invalid -> quote it and escape
-        ('"first"second"', '"""first""second"""'),
-        # valid quoted -> unchanged
-        ('"first""second"', '"first""second"'),
-    ],
-)
-def test_str_to_id_valid_values(input_val, expected_output):
-    definition = {
-        "definition_version": "1.1",
-        "env": {
-            "input_value": input_val,
-            "output_value": "<% fn.str_to_id(ctx.env.input_value) %>",
-        },
-    }
-
-    result = render_definition_template(definition, {}).project_context
-    env = result.get("ctx", {}).get("env", {})
-    assert env.get("output_value") == expected_output
-
-
-def test_str_to_id_with_no_args():
-    definition = {
-        "definition_version": "1.1",
-        "env": {
-            "value": "<% fn.str_to_id() %>",
-        },
-    }
-
-    with pytest.raises(InvalidTemplate) as err:
-        render_definition_template(definition, {})
-
-    assert "str_to_id requires at least 1 argument(s)" in err.value.message
-
-
-def test_str_to_id_with_two_args():
-    definition = {
-        "definition_version": "1.1",
-        "env": {
-            "value": "<% fn.str_to_id('a', 'b') %>",
-        },
-    }
-
-    with pytest.raises(InvalidTemplate) as err:
-        render_definition_template(definition, {})
-
-    assert "str_to_id supports at most 1 argument(s)" in err.value.message
-
-
-def test_str_to_id_with_non_string_arg():
-    definition = {
-        "definition_version": "1.1",
-        "env": {
-            "value": "<% fn.str_to_id(123) %>",
-        },
-    }
-
-    with pytest.raises(InvalidTemplate) as err:
-        render_definition_template(definition, {})
-
-    assert "str_to_id only accepts String values" in err.value.message
-
-
-@pytest.mark.parametrize(
-    "os_environ, expected_output",
-    [
-        ({"USER": "test_user"}, "test_user"),
-        ({"USERNAME": "test_user"}, "test_user"),
-        ({}, "unknown_user"),
-    ],
-)
-def test_get_username_valid_values(os_environ, expected_output):
-    definition = {
-        "definition_version": "1.1",
-        "env": {
-            "output_value": "<% fn.get_username() %>",
-        },
-    }
-
-    with mock.patch.dict(os.environ, os_environ, clear=True):
-        result = render_definition_template(definition, {}).project_context
-
-    env = result.get("ctx", {}).get("env", {})
-    assert env.get("output_value") == expected_output
-
-
-def test_get_username_with_args_should_fail():
-    definition = {
-        "definition_version": "1.1",
-        "env": {
-            "value": "<% fn.get_username('a') %>",
-        },
-    }
-
-    with pytest.raises(InvalidTemplate) as err:
-        render_definition_template(definition, {})
-
-    assert "get_username supports at most 0 argument(s)" in err.value.message
-
-
-@pytest.mark.parametrize(
-    "input_value, expected_output",
-    [
-        ("test_value", "test_value"),
-        (" T'EST_Va l.u-e" "", "test_value"),
-        ("", ""),
-        ('"some_id"', "some_id"),
-    ],
-)
-def test_clean_id_valid_values(input_value, expected_output):
-    definition = {
-        "definition_version": "1.1",
-        "env": {
-            "input_value": input_value,
-            "output_value": "<% fn.clean_id(ctx.env.input_value) %>",
-        },
-    }
-
-    result = render_definition_template(definition, {}).project_context
-
-    env = result.get("ctx", {}).get("env", {})
-    assert env.get("output_value") == expected_output
-
-
-def test_clean_id_with_no_args():
-    definition = {
-        "definition_version": "1.1",
-        "env": {
-            "value": "<% fn.clean_id() %>",
-        },
-    }
-
-    with pytest.raises(InvalidTemplate) as err:
-        render_definition_template(definition, {})
-
-    assert "clean_id requires at least 1 argument(s)" in err.value.message
-
-
-def test_clean_id_with_two_args():
-    definition = {
-        "definition_version": "1.1",
-        "env": {
-            "value": "<% fn.clean_id('a', 'b') %>",
-        },
-    }
-
-    with pytest.raises(InvalidTemplate) as err:
-        render_definition_template(definition, {})
-
-    assert "clean_id supports at most 1 argument(s)" in err.value.message
